[Test] 맛집 상세 정보 조회 통합 테스트 작성

+++ src/main/java/com/allclear/tastytrack/domain/review/repository/ReviewRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.allclear.tastytrack.domain.review.entity.Review;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
+
 	List<Review> findAllByRestaurantIdOrderByCreatedAtDesc(Integer restaurantId);
+
 }
